fix map-tools compare layers
