+++ addon/components/flexberry-maplayer.js
@@ -467,12 +467,25 @@ let FlexberryMaplayerComponent = Ember.Component.extend(
     isSubmenu: false,
 
     /**
+      Flag: indicates whether layer is group.
+
+      @property isSubmenu
+      @type boolean
+      @default false
+    */
+    isGroup: false,
+
+    /**
       Initializes DOM-related component's properties.
     */
     didInsertElement() {
+      if (this.get('layer.type') === 'group') {
+        this.set('isGroup', true);
+      }
+
       if (!this.get('readonly')) {
         let _this = this;
-        let $caption = Ember.$('label.flexberry-maplayer-caption-label');
+        let $caption = Ember.$('.ui.tab.treeview label.flexberry-maplayer-caption-label');
         if ($caption.length > 0) {
           $caption.hover(
             function() {
@@ -495,7 +508,115 @@ let FlexberryMaplayerComponent = Ember.Component.extend(
       }
     },
 
+    /**
+      Redefine L.Control.SideBySide._updateClip for work with layer and label layer.
+
+      @method updateClip
+    */
+    updateClip: function () {
+      let map = this.get('leafletMap');
+      let sbs = this.get('sideBySide');
+      var nw = map.containerPointToLayerPoint([0, 0]);
+      var se = map.containerPointToLayerPoint(map.getSize());
+      var clipX = nw.x + sbs.getPosition();
+      var dividerX = sbs.getPosition();
+      sbs._divider.style.left = dividerX + 'px';
+      var clipLeft = 'rect(' + [nw.y, clipX, se.y, nw.x].join('px,') + 'px)';
+      var clipRight = 'rect(' + [nw.y, se.x, se.y, clipX].join('px,') + 'px)';
+      sbs._leftLayers.forEach(function (layer) {
+        if (!Ember.isNone(layer) && !Ember.isNone(layer.getContainer)) {
+          layer.getContainer().style.clip = clipLeft;
+        }
+      });
+      sbs._rightLayers.forEach(function (layer) {
+        if (!Ember.isNone(layer) && !Ember.isNone(layer.getContainer)) {
+          layer.getContainer().style.clip = clipRight;
+        }
+      });
+    },
+
     actions: {
+      /**
+        Handles click on checkbox.
+        @method action.onChange
+        @param {Object} e event args.
+      */
+      onChange(e) {
+        let layer = this.get('layer');
+        let sbs = this.get('sideBySide');
+        Ember.set(sbs, 'baseUpdateClip', sbs._updateClip);
+        if (e.newValue === false) {
+          sbs.off('dividermove', this.updateClip, this);
+          if (this.get('side') === 'Left') {
+            sbs.setLeftLayers(null);
+            this.get('leftLayer._leafletObject').remove();
+            if (layer.get('settingsAsObject.labelSettings.signMapObjects')) {
+              this.get('leftLayer._leafletObject._labelsLayer').remove();
+            }
+
+            this.set('leftLayer.side', null);
+            this.set('leftLayer.visibility', false);
+            this.set('leftLayer', null);
+          } else {
+            sbs.setRightLayers(null);
+            this.get('rightLayer._leafletObject').remove();
+            if (layer.get('settingsAsObject.labelSettings.signMapObjects')) {
+              this.get('rightLayer._leafletObject._labelsLayer').remove();
+            }
+
+            this.set('rightLayer.side', null);
+            this.set('rightLayer.visibility', false);
+            this.set('rightLayer', null);
+          }
+        } else {
+          Ember.set(layer, 'visibility', true);
+          let map = this.get('leafletMap');
+          sbs._updateClip = this.get('updateClip').bind(this);
+          sbs.on('dividermove', this.updateClip, this);
+
+          if (this.get('side') === 'Left') {
+            if (this.get('leftLayer') !== null) {
+              sbs.setLeftLayers(null);
+              this.get('leftLayer._leafletObject').remove();
+              this.set('leftLayer.visibility', false);
+              this.set('leftLayer.side', null);
+            }
+
+            let leafletObject = Ember.get(layer, '_leafletObject').addTo(map);
+            let left = [];
+            left.push(leafletObject);
+            if (layer.get('settingsAsObject.labelSettings.signMapObjects')) {
+              left.push(leafletObject._labelsLayer.addTo(map));
+            }
+
+            Ember.set(layer, 'side', 'Left');
+            this.set('leftLayer', layer);
+            sbs.setLeftLayers(left);
+          }
+
+          if (this.get('side') === 'Right')  {
+            if (this.get('rightLayer') !== null) {
+              sbs.setRightLayers(null);
+              this.get('rightLayer._leafletObject').remove();
+              this.set('rightLayer.visibility', false);
+              this.set('rightLayer.side', null);
+            }
+
+            let leafletObject = Ember.get(layer, '_leafletObject').addTo(map);
+            let right = [];
+            right.push(leafletObject);
+            if (layer.get('settingsAsObject.labelSettings.signMapObjects')) {
+              right.push(leafletObject._labelsLayer.addTo(map));
+            }
+
+            Ember.set(layer, 'side', 'Right');
+            this.set('rightLayer', layer);
+            sbs.setRightLayers(right);
+          }
+        }
+        this.sendAction('onChangeLayer', this.get('leftLayer'), this.get('rightLayer'));
+      },
+
       /**
         Show\hide submenu
