+++ tests/dummy/app/locales/en/forms/map.js
@@ -7,6 +7,7 @@ export default {
   'favoritesbuttontooltip': 'Show favorites',
   'createobjectbuttontooltip': 'Show object creation',
   'createoreditobjectbuttontooltip': 'Show object creation/editing',
+  'comparebuttontooltip': 'Compare layers',
   'xCaption': 'Longitude',
   'yCaption': 'Latitude',
 
@@ -31,6 +32,9 @@ export default {
     },
     'createOrEditObject': {
       'caption': 'Create/Edit object'
+    },
+    'compare': {
+      'caption': 'Compare layers'
     }
   }
 };
